Refactored code to remove redundancy between AD and HSXL code
