+++ animatediff/model_utils.py
@@ -13,38 +13,6 @@ import folder_paths
 from comfy.model_base import SDXL, BaseModel
 from comfy.model_patcher import ModelPatcher
 from comfy.model_management import xformers_enabled
-from abc import ABC, abstractmethod
-
-
-from .motion_lora import MotionLoRAInfo
-
-
-class InjectorVersion:
-    V1_V2 = "v1/v2"
-    HOTSHOTXL_V1 = "HSXL v1"
-
-
-class GenericMotionWrapper(nn.Module, ABC):
-    def __init__(self, mm_hash: str, mm_name: str, loras: list[MotionLoRAInfo]):
-        super().__init__()
-        self.down_blocks: nn.ModuleList = None
-        self.up_blocks: nn.ModuleList = None
-        self.mid_block = None
-        self.mm_hash = mm_hash
-        self.mm_name = mm_name
-        self.version = "FILLTHISIN"
-        self.injector_version = "VERYIMPORTANT_FILLTHISIN"
-        self.AD_video_length: int = 0
-        self.loras = loras
-
-    def has_loras(self) -> bool:
-        # TODO: fix this to return False if has an empty list as well
-        # but only after implementing a fix for lowvram loading
-        return self.loras is not None
-
-    @abstractmethod
-    def set_video_length(self, video_length: int):
-        pass
 
 
 class IsChangedHelper:
